Added Userge antiSpam Api to antispam.py (#338)
* Added Userge antiSpam Api to antispam.py

* added lib to requirements

* pep8

* Update antispam.py

* fix probability to spam logs

* bug fix
