+++ config.env.sample
@@ -88,6 +88,10 @@ OCR_SPACE_API_KEY=""
 WEATHER_DEFCITY=""
 
 
+# Userge AntiSpam API get it from @UsergeAntiSpamBot in Telegram
+USERGE_ANTISPAM_API=""
+
+
 # SpamWatch API get it from @SpamWatch in Telegram
 SPAM_WATCH_API=""
