+++ userge/config.py
@@ -46,6 +46,7 @@ class Config:
     CUSTOM_PACK_NAME = os.environ.get("CUSTOM_PACK_NAME")
     UPSTREAM_REPO = os.environ.get("UPSTREAM_REPO")
     UPSTREAM_REMOTE = os.environ.get("UPSTREAM_REMOTE")
+    USERGE_ANTISPAM_API = os.environ.get("USERGE_ANTISPAM_API")
     SPAM_WATCH_API = os.environ.get("SPAM_WATCH_API")
     CURRENCY_API = os.environ.get("CURRENCY_API")
     OCR_SPACE_API_KEY = os.environ.get("OCR_SPACE_API_KEY")
